# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EdeQR LMS API",
        "description": "bridge digital LMS projesi için REST API dokümantasyonu",
        "contact": {
            "email": "ahmet@aciksari.net"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/schools/{school}/manager/sessions": {
            "get": {
                "tags": [
                    "School Session Management"
                ],
                "summary": "Okuldaki tüm ders oturumlarını listeler (Teacher ise kendi oturumları)",
                "operationId": "listSchoolSessions",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "description": "Okul ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SchoolSession"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz Erişim"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "School Session Management"
                ],
                "summary": "Yeni bir ders oturumu kaydı oluşturur",
                "operationId": "storeSchoolSession",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSchoolSessionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SchoolSession"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz Erişim"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/schools/{school}/manager/sessions/{session}": {
            "get": {
                "tags": [
                    "School Session Management"
                ],
                "summary": "Belirli bir ders oturumunun detaylarını getirir",
                "operationId": "showSchoolSession",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "description": "Okul ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "session",
                        "in": "path",
                        "description": "Ders Oturumu ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SchoolSession"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oturum bulunamadı veya okula ait değil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "School Session Management"
                ],
                "summary": "Belirli bir ders oturumunu günceller",
                "operationId": "updateSchoolSession",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "description": "Okul ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "session",
                        "in": "path",
                        "description": "Ders Oturumu ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSchoolSessionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SchoolSession"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oturum bulunamadı veya okula ait değil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "School Session Management"
                ],
                "summary": "Belirli bir ders oturumunu siler",
                "operationId": "deleteSchoolSession",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "description": "Okul ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "session",
                        "in": "path",
                        "description": "Ders Oturumu ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Başarılı (İçerik Yok)"
                    },
                    "404": {
                        "description": "Oturum bulunamadı veya okula ait değil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/me/teacher/getprofilesettings": {
            "get": {
                "tags": [
                    "TeacherProfileSettings"
                ],
                "summary": "Öğretmen profil ayarlarını getir",
                "operationId": "0a171f0f8c6aa11024768e8d3bb704bc",
                "responses": {
                    "200": {
                        "description": "Profil ayarları bulundu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeacherProfile"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim"
                    },
                    "404": {
                        "description": "Profil bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me/teacher/updateprofilesettings": {
            "put": {
                "tags": [
                    "TeacherProfileSettings"
                ],
                "summary": "Öğretmen profil ayarlarını güncelle",
                "operationId": "0fef0a47f1903c0f2e63daa0c928369e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TeacherUpdateProfileSettingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeacherProfile"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim"
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TeacherUpdateProfileSettingRequest": {
                "title": "Teacher Update Profile Settings",
                "description": "Öğretmenin profil ayarlarını güncellemek için kullanılan alanlar.",
                "required": [],
                "properties": {
                    "phone": {
                        "type": "string",
                        "example": "+905551234567",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "İzmir, Türkiye",
                        "nullable": true
                    },
                    "schoolId": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "branch_id": {
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "img_path": {
                        "type": "string",
                        "example": "/storage/teachers/avatars/5.png",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "example": "female",
                        "nullable": true
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-02-14",
                        "nullable": true
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-09-01",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "STEM kulübü sorumlusu",
                        "nullable": true
                    },
                    "color_code": {
                        "type": "string",
                        "example": "#ff9800",
                        "nullable": true
                    },
                    "emergency_contact_name": {
                        "type": "string",
                        "example": "Ahmet Öğretmen",
                        "nullable": true
                    },
                    "emergency_contact_phone": {
                        "type": "string",
                        "example": "+905559998877",
                        "nullable": true
                    },
                    "emergency_contact_relationship": {
                        "type": "string",
                        "example": "Kardeşi",
                        "nullable": true
                    },
                    "emergency_contact_description": {
                        "type": "string",
                        "example": "Hafta içi 09:00-18:00 arası ulaşılabilir.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StoreSchoolSessionRequest": {
                "title": "Ders Oturumu Oluşturma İsteği",
                "required": [
                    "class_id",
                    "subject_id",
                    "teacher_id",
                    "session_date",
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "class_id": {
                        "description": "Dersin yapılacağı Sınıf ID",
                        "type": "integer",
                        "example": 5
                    },
                    "subject_id": {
                        "description": "Dersin (Subject) ID",
                        "type": "integer",
                        "example": 12
                    },
                    "teacher_id": {
                        "description": "Dersi verecek Öğretmen (User) ID",
                        "type": "integer",
                        "example": 205
                    },
                    "session_date": {
                        "description": "Oturumun yapılacağı tarih",
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-25"
                    },
                    "start_time": {
                        "description": "Dersin başlangıç saati (HH:MM:SS)",
                        "type": "string",
                        "format": "time",
                        "example": "09:00:00"
                    },
                    "end_time": {
                        "description": "Dersin bitiş saati (HH:MM:SS)",
                        "type": "string",
                        "format": "time",
                        "example": "10:30:00"
                    },
                    "location": {
                        "description": "Dersin yapılacağı yer",
                        "type": "string",
                        "example": "B-Blok 301 Nolu Sınıf",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateSchoolSessionRequest": {
                "title": "Ders Oturumu Güncelleme İsteği",
                "properties": {
                    "teacher_id": {
                        "description": "Dersi verecek Öğretmen (User) ID",
                        "type": "integer",
                        "example": 206
                    },
                    "start_time": {
                        "description": "Dersin başlangıç saati (HH:MM:SS)",
                        "type": "string",
                        "format": "time",
                        "example": "10:00:00"
                    }
                },
                "type": "object"
            },
...
            "SchoolSession": {
                "title": "School Session Model",
                "description": "Belirli bir sınıf, öğretmen, ders, gün ve saate ait ders oturumu.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 15
                    },
                    "school_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "class_model_id": {
                        "description": "Dersin verildiği sınıf",
                        "type": "integer",
                        "example": 2
                    },
                    "teacher_id": {
                        "description": "Dersi veren öğretmenin User ID'si",
                        "type": "integer",
                        "example": 5
                    },
                    "subject_id": {
                        "description": "Dersin konusu",
                        "type": "integer",
                        "example": 3
                    },
                    "school_week_id": {
                        "description": "Dersin gerçekleştiği haftanın kaydı",
                        "type": "integer",
                        "example": 10
                    },
                    "school_day_id": {
                        "description": "Dersin gerçekleştiği günün kaydı",
                        "type": "integer",
                        "example": 2
                    },
                    "start_time": {
                        "type": "string",
                        "format": "time",
                        "example": "09:00:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "time",
                        "example": "10:00:00"
                    },
                    "date": {
                        "description": "Dersin gerçekleştiği somut tarih (otomatik hesaplanabilir)",
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-25",
                        "nullable": true
                    },
                    "is_cancelled": {
                        "description": "Dersin iptal edilip edilmediği",
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "TeacherProfile": {
                "title": "TeacherProfile",
                "description": "Öğretmen profil modeli",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "phone": {
                        "type": "string",
                        "example": "+905551234567",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "Ankara, Türkiye",
                        "nullable": true
                    },
                    "schoolId": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "branch_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "img_path": {
                        "type": "string",
                        "example": "/storage/teachers/1.png",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "example": "female",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "example": "1992-03-10",
                        "nullable": true
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-09-01",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Matematik öğretmeni",
                        "nullable": true
                    },
                    "color_code": {
                        "type": "string",
                        "example": "#3490dc",
                        "nullable": true
                    },
                    "emergency_contact_name": {
                        "type": "string",
                        "example": "Veli Öğretmen",
                        "nullable": true
                    },
                    "emergency_contact_phone": {
                        "type": "string",
                        "example": "+905559998877",
                        "nullable": true
                    },
                    "emergency_contact_relationship": {
                        "type": "string",
                        "example": "Eşi",
                        "nullable": true
                    },
                    "emergency_contact_description": {
                        "type": "string",
                        "example": "Hafta içi ulaşılabilir.",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-21T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-21T12:00:00Z"
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User Model",
                "description": "Kullanıcı modelini temsil eder.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ahmet Açısarı"
                    },
                    "email": {
                        "type": "string",
                        "example": "ahmet@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT veya Sanctum tokenini \"Bearer {token}\" formatında girin"
            },
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "1": {
            "bearerAuth": []
        }
    }
}